#653: rdma: factor out common code in tests
